Extract reusable QIO Jacobi optimizer
