+++ cpp/src/qdk/chemistry/algorithms/microsoft/CMakeLists.txt
@@ -10,6 +10,7 @@ target_sources(chemistry PRIVATE
     localization/pipek_mezey.cpp
     localization/active_space_qio.cpp
     localization/vvhv.cpp
+    qio/jacobi_optimizer.cpp
     hamiltonian.cpp
     cholesky_hamiltonian.cpp
     macis_asci.cpp
