+++ cpp/src/qdk/chemistry/algorithms/microsoft/localization/active_space_qio.cpp
@@ -6,12 +6,9 @@
 
 #include <Eigen/Dense>
 #include <algorithm>
-#include <array>
 #include <cmath>
 #include <cstdint>
-#include <numbers>
 #include <optional>
-#include <qdk/chemistry/data/orbital_entropy.hpp>
 #include <qdk/chemistry/data/symmetry/spin_channel_indices.hpp>
 #include <qdk/chemistry/data/symmetry/symmetry_blocked_index_set.hpp>
 #include <qdk/chemistry/data/symmetry/symmetry_blocked_tensor.hpp>
@@ -22,284 +19,9 @@
 #include <variant>
 #include <vector>
 
+#include "../qio/jacobi_optimizer.hpp"
+
 namespace qdk::chemistry::algorithms::microsoft {
-
-namespace detail {
-
-/**
- * @brief Flat row-major offset into an (dim x dim x dim x dim) tensor.
- *
- * @param dim Dimension of each of the four tensor axes.
- * @param i First axis index.
- * @param j Second axis index.
- * @param k Third axis index.
- * @param l Fourth axis index.
- * @return The flattened row-major offset.
- */
-inline std::size_t flat_index(std::size_t dim, std::size_t i, std::size_t j,
-                              std::size_t k, std::size_t l) {
-  return ((i * dim + j) * dim + k) * dim + l;
-}
-
-/**
- * @brief Apply an in-place Givens rotation to one axis of a rank-4 2-RDM
- * tensor.
- *
- * The transform on the active axis is v_i <- c*v_i + s*v_j and
- * v_j <- -s*v_i + c*v_j, with all four axes of dimension @p dim.
- *
- * @param rdm_aabb Flattened (dim^4) alpha-beta 2-RDM block, modified in place.
- * @param dim Dimension of each tensor axis.
- * @param axis Axis (0-3) to which the rotation is applied.
- * @param i First orbital of the rotation plane.
- * @param j Second orbital of the rotation plane.
- * @param cos_theta Cosine of the rotation angle.
- * @param sin_theta Sine of the rotation angle.
- */
-void rotate_two_rdm_axis(std::vector<double>& rdm_aabb, std::size_t dim,
-                         int axis, std::size_t i, std::size_t j,
-                         double cos_theta, double sin_theta) {
-  const std::array<std::size_t, 4> strides{dim * dim * dim, dim * dim, dim, 1};
-  const std::size_t axis_stride = strides[axis];
-  const std::size_t axis_span = dim * axis_stride;
-  const std::size_t block_count = dim * dim * dim;
-
-#pragma omp parallel for schedule(static)
-  for (std::size_t block = 0; block < block_count; ++block) {
-    const std::size_t base =
-        (block / axis_stride) * axis_span + block % axis_stride;
-    double& elem_i = rdm_aabb[base + i * axis_stride];
-    double& elem_j = rdm_aabb[base + j * axis_stride];
-    const double val_i = elem_i;
-    const double val_j = elem_j;
-    elem_i = cos_theta * val_i + sin_theta * val_j;
-    elem_j = -sin_theta * val_i + cos_theta * val_j;
-  }
-}
-
-/**
- * @brief Apply a Givens rotation to all four axes of a 2-RDM tensor.
- *
- * @param rdm_aabb Flattened (dim^4) alpha-beta 2-RDM block, modified in place.
- * @param dim Dimension of each tensor axis.
- * @param i First orbital of the rotation plane.
- * @param j Second orbital of the rotation plane.
- * @param cos_theta Cosine of the rotation angle.
- * @param sin_theta Sine of the rotation angle.
- */
-void rotate_two_rdm(std::vector<double>& rdm_aabb, std::size_t dim,
-                    std::size_t i, std::size_t j, double cos_theta,
-                    double sin_theta) {
-  for (int axis = 0; axis < 4; ++axis) {
-    rotate_two_rdm_axis(rdm_aabb, dim, axis, i, j, cos_theta, sin_theta);
-  }
-}
-
-/**
- * @brief Apply a Givens rotation to both axes of a (symmetric) 1-RDM.
- *
- * @param rdm The (dim x dim) 1-RDM matrix, modified in place.
- * @param i First orbital of the rotation plane.
- * @param j Second orbital of the rotation plane.
- * @param cos_theta Cosine of the rotation angle.
- * @param sin_theta Sine of the rotation angle.
- */
-void rotate_one_rdm(Eigen::MatrixXd& rdm, std::size_t i, std::size_t j,
-                    double cos_theta, double sin_theta) {
-  const Eigen::Index ii = static_cast<Eigen::Index>(i);
-  const Eigen::Index jj = static_cast<Eigen::Index>(j);
-  // Left rotation: mix rows i and j (iterate over all columns p).
-  for (Eigen::Index p = 0; p < rdm.cols(); ++p) {
-    const double row_i = rdm(ii, p);
-    const double row_j = rdm(jj, p);
-    rdm(ii, p) = cos_theta * row_i + sin_theta * row_j;
-    rdm(jj, p) = -sin_theta * row_i + cos_theta * row_j;
-  }
-  // Right rotation: mix columns i and j (iterate over all rows p).
-  for (Eigen::Index p = 0; p < rdm.rows(); ++p) {
-    const double col_i = rdm(p, ii);
-    const double col_j = rdm(p, jj);
-    rdm(p, ii) = cos_theta * col_i + sin_theta * col_j;
-    rdm(p, jj) = -sin_theta * col_i + cos_theta * col_j;
-  }
-}
-
-/**
- * @brief Total single-orbital entropy sum F_QI = sum_i S(rho_i).
- *
- * @param rdm_alpha Active alpha 1-RDM (dim x dim).
- * @param rdm_beta Active beta 1-RDM (dim x dim).
- * @param rdm_aabb Flattened (dim^4) alpha-beta 2-RDM block.
- * @param dim Active-space dimension.
- * @return The summed single-orbital entropy over all active orbitals.
- */
-double total_single_orbital_entropy(const Eigen::MatrixXd& rdm_alpha,
-                                    const Eigen::MatrixXd& rdm_beta,
-                                    const std::vector<double>& rdm_aabb,
-                                    std::size_t dim) {
-  double entropy = 0.0;
-  for (std::size_t i = 0; i < dim; ++i) {
-    const Eigen::Index ii = static_cast<Eigen::Index>(i);
-    entropy += data::detail::single_orbital_entropy(
-        rdm_alpha(ii, ii), rdm_beta(ii, ii),
-        rdm_aabb[flat_index(dim, i, i, i, i)]);
-  }
-  return entropy;
-}
-
-struct RotationOptimizationResult {
-  Eigen::MatrixXd rotation;
-  Eigen::MatrixXd rdm_alpha;
-  Eigen::MatrixXd rdm_beta;
-};
-
-/**
- * @brief Minimize the single-orbital entropy sum by gradient-free Jacobi
- * sweeps.
- *
- * Each active orbital pair (i, j) is rotated by the angle that minimizes the
- * two entropy terms that change, located by a coarse-then-fine 1-D scan. The
- * plane rotation is applied to the cached RDMs and accumulated into a single
- * unitary, which is returned.
- *
- * @param rdm_alpha Active alpha 1-RDM (dim x dim).
- * @param rdm_beta Active beta 1-RDM (dim x dim).
- * @param rdm_aabb Flattened (dim^4) alpha-beta 2-RDM block.
- * @param dim Active-space dimension.
- * @param max_cycles Maximum number of Jacobi sweeps.
- * @param convergence_tolerance Sweep-to-sweep entropy-sum change to stop at.
- * @param coarse_angle_step Coarse angle-scan spacing (radians) over [0, pi/2).
- * @param fine_samples Number of samples in the fine refinement scan.
- * @param improvement_tolerance Minimum entropy decrease to accept a rotation.
- * @return The accumulated rotation and rotated spin-resolved 1-RDMs.
- */
-RotationOptimizationResult optimize_rotation(
-    Eigen::MatrixXd rdm_alpha, Eigen::MatrixXd rdm_beta,
-    std::vector<double> rdm_aabb, std::size_t dim, std::size_t max_cycles,
-    double convergence_tolerance, double coarse_angle_step, int fine_samples,
-    double improvement_tolerance) {
-  Eigen::MatrixXd rotation = Eigen::MatrixXd::Identity(
-      static_cast<Eigen::Index>(dim), static_cast<Eigen::Index>(dim));
-  if (dim < 2) {
-    return {std::move(rotation), std::move(rdm_alpha), std::move(rdm_beta)};
-  }
-
-  double entropy_prev =
-      total_single_orbital_entropy(rdm_alpha, rdm_beta, rdm_aabb, dim);
-  constexpr double angle_period = std::numbers::pi / 2.0;
-  for (std::size_t cycle = 0; cycle < max_cycles; ++cycle) {
-    for (std::size_t i = 0; i < dim; ++i) {
-      for (std::size_t j = i + 1; j < dim; ++j) {
-        const Eigen::Index ii = static_cast<Eigen::Index>(i);
-        const Eigen::Index jj = static_cast<Eigen::Index>(j);
-        const std::array<std::size_t, 2> pair{i, j};
-        const double a_ii = rdm_alpha(ii, ii), a_ij = rdm_alpha(ii, jj),
-                     a_jj = rdm_alpha(jj, jj);
-        const double b_ii = rdm_beta(ii, ii), b_ij = rdm_beta(ii, jj),
-                     b_jj = rdm_beta(jj, jj);
-
-        std::array<double, 5> double_occupation_coefficients{};
-        for (std::size_t a = 0; a < 2; ++a) {
-          for (std::size_t b = 0; b < 2; ++b) {
-            for (std::size_t c = 0; c < 2; ++c) {
-              for (std::size_t d = 0; d < 2; ++d) {
-                double_occupation_coefficients[a + b + c + d] +=
-                    rdm_aabb[flat_index(dim, pair[a], pair[b], pair[c],
-                                        pair[d])];
-              }
-            }
-          }
-        }
-
-        auto double_occupation = [&](double w0, double w1) {
-          const double w0_squared = w0 * w0;
-          const double w1_squared = w1 * w1;
-          return double_occupation_coefficients[0] * w0_squared * w0_squared +
-                 double_occupation_coefficients[1] * w0_squared * w0 * w1 +
-                 double_occupation_coefficients[2] * w0_squared * w1_squared +
-                 double_occupation_coefficients[3] * w0 * w1_squared * w1 +
-                 double_occupation_coefficients[4] * w1_squared * w1_squared;
-        };
-
-        // Pair entropy after rotating by theta: i' = (c, s), j' = (-s, c).
-        auto pair_entropy = [&](double theta) {
-          const double cos_theta = std::cos(theta);
-          const double sin_theta = std::sin(theta);
-          const double occ_alpha_i = cos_theta * cos_theta * a_ii +
-                                     2.0 * cos_theta * sin_theta * a_ij +
-                                     sin_theta * sin_theta * a_jj;
-          const double occ_beta_i = cos_theta * cos_theta * b_ii +
-                                    2.0 * cos_theta * sin_theta * b_ij +
-                                    sin_theta * sin_theta * b_jj;
-          const double occ_alpha_j = sin_theta * sin_theta * a_ii -
-                                     2.0 * cos_theta * sin_theta * a_ij +
-                                     cos_theta * cos_theta * a_jj;
-          const double occ_beta_j = sin_theta * sin_theta * b_ii -
-                                    2.0 * cos_theta * sin_theta * b_ij +
-                                    cos_theta * cos_theta * b_jj;
-          const double double_occ_i = double_occupation(cos_theta, sin_theta);
-          const double double_occ_j = double_occupation(-sin_theta, cos_theta);
-          return data::detail::single_orbital_entropy(occ_alpha_i, occ_beta_i,
-                                                      double_occ_i) +
-                 data::detail::single_orbital_entropy(occ_alpha_j, occ_beta_j,
-                                                      double_occ_j);
-        };
-
-        const double entropy_current = pair_entropy(0.0);
-        double best_theta = 0.0;
-        double best_entropy = entropy_current;
-        for (double theta = 0.0; theta < angle_period;
-             theta += coarse_angle_step) {
-          const double value = pair_entropy(theta);
-          if (value < best_entropy) {
-            best_entropy = value;
-            best_theta = theta;
-          }
-        }
-        const double theta_lo = best_theta - coarse_angle_step;
-        const double theta_hi = best_theta + coarse_angle_step;
-        for (int k = 0; k < fine_samples; ++k) {
-          const double theta =
-              theta_lo +
-              (theta_hi - theta_lo) * k / static_cast<double>(fine_samples - 1);
-          const double value = pair_entropy(theta);
-          if (value < best_entropy) {
-            best_entropy = value;
-            best_theta = theta;
-          }
-        }
-        best_theta = std::fmod(best_theta, angle_period);
-        if (best_theta < 0.0) {
-          best_theta += angle_period;
-        }
-
-        if (best_entropy < entropy_current - improvement_tolerance) {
-          const double cos_theta = std::cos(best_theta);
-          const double sin_theta = std::sin(best_theta);
-          rotate_one_rdm(rdm_alpha, i, j, cos_theta, sin_theta);
-          rotate_one_rdm(rdm_beta, i, j, cos_theta, sin_theta);
-          rotate_two_rdm(rdm_aabb, dim, i, j, cos_theta, sin_theta);
-          // Accumulate U <- U * G(i, j; c, s) (in-place column update).
-          for (Eigen::Index p = 0; p < rotation.rows(); ++p) {
-            const double rot_i = rotation(p, ii);
-            const double rot_j = rotation(p, jj);
-            rotation(p, ii) = cos_theta * rot_i + sin_theta * rot_j;
-            rotation(p, jj) = -sin_theta * rot_i + cos_theta * rot_j;
-          }
-        }
-      }
-    }
-    const double entropy_now =
-        total_single_orbital_entropy(rdm_alpha, rdm_beta, rdm_aabb, dim);
-    if (std::abs(entropy_prev - entropy_now) < convergence_tolerance) {
-      break;
-    }
-    entropy_prev = entropy_now;
-  }
-  return {std::move(rotation), std::move(rdm_alpha), std::move(rdm_beta)};
-}
-
-}  // namespace detail
 
 std::shared_ptr<data::Wavefunction> ActiveSpaceQIOLocalizer::_run_impl(
     std::shared_ptr<data::Wavefunction> wavefunction,
@@ -439,7 +161,7 @@ std::shared_ptr<data::Wavefunction> ActiveSpaceQIOLocalizer::_run_impl(
   require_finite("convergence_tolerance", convergence_tolerance);
   require_finite("coarse_angle_step", coarse_angle_step);
   require_finite("improvement_tolerance", improvement_tolerance);
-  auto optimization = detail::optimize_rotation(
+  auto optimization = qio::detail::optimize_rotation(
       std::move(rdm_alpha), std::move(rdm_beta), std::move(rdm_aabb_flat), n,
       max_cycles, convergence_tolerance, coarse_angle_step, fine_samples,
       improvement_tolerance);
